third_party/WebKit/LayoutTests/webaudio/AudioParam/worklet-warnings.html

<!doctype html>
<html>
  <head>
    <title>
      Test AudioWorklet AudioParam Warning Messages
    </title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script src="../resources/audit-util.js"></script>
    <script src="../resources/audit.js"></script>
  </head>
  <body>
    <script id="layout-test-code">
      let audit = Audit.createTaskRunner();

      // A simple noise generator.  The actual processor doesn't matter; it just
      // needs to have an AudioParam with finite nominal range where we can
      // automate it with values outside the nominal range.
      let workletScript = 
`
class NoiseGenerator extends AudioWorkletProcessor {
  static get parameterDescriptors() {
    return [{name: 'amplitude', defaultValue: 0.25, minValue: 0, maxValue: 1}];
  }

  constructor(options) {
    super(options);
  }

  process(inputs, outputs, parameters) {
    let output = outputs[0];
    let amplitude = parameters.amplitude;
    for (let channel = 0; channel < output.length; ++channel) {
      let outputChannel = output[channel];
      for (let i = 0; i < outputChannel.length; ++i) {
        outputChannel[i] = 2 * (Math.random() - 0.5) * amplitude[i];
      }
    }

    return true;
  }
}

registerProcessor('noise-generator', NoiseGenerator);
`;

      const url = window.URL.createObjectURL(
          new Blob([workletScript], {type: 'text/javascript'}));

      audit.define(
          {
            label: 'AudioWorklet AudioParam',
            description: 'Generate warnings if outside nominal range'
          },
          (task, should) => {window.audioWorklet.addModule(url).then(() => {
            // Any reasonable sample rate will work
            let sampleRate = 16000;
            let renderTime = 1;
            let renderLength = renderTime * sampleRate;
            let context = new OfflineAudioContext(1, renderLength, sampleRate);
            let noiseGenerator =
                new AudioWorkletNode(context, 'noise-generator');
            noiseGenerator.connect(context.destination);
            let param = noiseGenerator.parameters.get('amplitude');
            // Set the value inside the nominal range; no warning should be
            // generated.
            param.value = .1;
            // Set the value outside the nominal range to generate a warning.
            param.value = 99;

            // Set up automation outside the nominal range to generate a
            // warning.
            param.setValueAtTime(-1, renderTime / 4);
            param.linearRampToValueAtTime(5, renderTime);

            // Render; we don't care what the generated result is.
            context.startRendering()
                .then(() => {
                  should(true, 'Rendering succeeded').beTrue();
                })
                .then(() => task.done());
          })});

      audit.run();
    </script>
  </body>
</html>
